+++ SolarSystemProject/SolarSystemProject/Sun.h
@@ -9,7 +9,7 @@ private:
 	std::string name;
 
 public:
-	Sun(int size, float positions[2],float initialVelocity, int temperature,std::string color,std::string name);
+	Sun(int size, int positions[2],float initialVelocity, int temperature,std::string color,std::string name);
 	bool checkTemperature() override;
 	void increaseTemperature() override;
     int getTemperature() const;
